#2392 starting off the split and rename process
