+++ src/psyclone/tests/dynamo0p3_basis_test.py
@@ -43,8 +43,8 @@ import pytest
 import fparser
 from fparser import api as fpapi
 from psyclone.configuration import Config
-from psyclone.domain.lfric import LFRicConstants, LFRicKern
-from psyclone.dynamo0p3 import DynKernMetadata, DynBasisFunctions
+from psyclone.domain.lfric import LFRicConstants, LFRicKern, LFRicKernMetadata
+from psyclone.dynamo0p3 import DynBasisFunctions
 from psyclone.f2pygen import ModuleGen
 from psyclone.parse.algorithm import parse
 from psyclone.parse.utils import ParseError
@@ -96,7 +96,7 @@ def test_eval_mdata():
     ''' Check that we recognise "evaluator" as a valid gh_shape '''
     fparser.logging.disable(fparser.logging.CRITICAL)
     ast = fpapi.parse(CODE, ignore_comments=False)
-    dkm = DynKernMetadata(ast, name="testkern_eval_type")
+    dkm = LFRicKernMetadata(ast, name="testkern_eval_type")
     assert dkm.get_integer_variable('gh_shape') == 'gh_evaluator'
 
 
@@ -107,13 +107,13 @@ def test_multi_updated_arg():
     # Change the access of the read-only argument
     code = CODE.replace("GH_READ", "GH_INC", 1)
     ast = fpapi.parse(code, ignore_comments=False)
-    dkm = DynKernMetadata(ast, name="testkern_eval_type")
+    dkm = LFRicKernMetadata(ast, name="testkern_eval_type")
     # Evaluator targets list remains unchanged
     assert dkm._eval_targets == ['w0', 'w1']
     # Change the gh_shape element to specify quadrature and then test again
     qr_code = code.replace("gh_evaluator", "gh_quadrature_xyoz")
     ast = fpapi.parse(qr_code, ignore_comments=False)
-    dkm = DynKernMetadata(ast, name="testkern_eval_type")
+    dkm = LFRicKernMetadata(ast, name="testkern_eval_type")
     assert dkm.get_integer_variable('gh_shape') == "gh_quadrature_xyoz"
 
 
@@ -121,7 +121,7 @@ def test_eval_targets():
     ''' Check that we can specify multiple evaluator targets using
     the gh_evaluator_targets meta-data entry. '''
     ast = fpapi.parse(CODE, ignore_comments=False)
-    dkm = DynKernMetadata(ast, name="testkern_eval_type")
+    dkm = LFRicKernMetadata(ast, name="testkern_eval_type")
     assert dkm._eval_targets == ["w0", "w1"]
 
 
@@ -132,7 +132,7 @@ def test_eval_targets_err():
     code = CODE.replace("gh_evaluator\n", "gh_quadrature_xyoz\n")
     ast = fpapi.parse(code, ignore_comments=False)
     with pytest.raises(ParseError) as err:
-        _ = DynKernMetadata(ast, name="testkern_eval_type")
+        _ = LFRicKernMetadata(ast, name="testkern_eval_type")
     assert ("specifies gh_evaluator_targets (['w0', 'w1']) but does not need "
             "an evaluator because gh_shape=['gh_quadrature_xyoz']"
             in str(err.value))
@@ -146,7 +146,7 @@ def test_eval_targets_err():
                         "")
     ast = fpapi.parse(code, ignore_comments=False)
     with pytest.raises(ParseError) as err:
-        _ = DynKernMetadata(ast, name="testkern_eval_type")
+        _ = LFRicKernMetadata(ast, name="testkern_eval_type")
     assert ("specifies gh_evaluator_targets (['w0', 'w1']) but does not need "
             "an evaluator because no basis or differential basis functions "
             "are required" in str(err.value))
@@ -158,7 +158,7 @@ def test_eval_targets_wrong_space():
     code = CODE.replace("[W0, W1]", "[W0, W3]")
     ast = fpapi.parse(code, ignore_comments=False)
     with pytest.raises(ParseError) as err:
-        _ = DynKernMetadata(ast, name="testkern_eval_type")
+        _ = LFRicKernMetadata(ast, name="testkern_eval_type")
     assert ("specifies that an evaluator is required on 'w3' but does not "
             "have an argument on this space" in str(err.value))
 
@@ -173,14 +173,14 @@ def test_eval_targets_op_space():
     code = code.replace("[W0, W1]", "[W0, W3]")
     ast = fpapi.parse(code, ignore_comments=False)
     with pytest.raises(ParseError) as err:
-        _ = DynKernMetadata(ast, name="testkern_eval_type")
+        _ = LFRicKernMetadata(ast, name="testkern_eval_type")
     assert ("specifies that an evaluator is required on 'w3' but does not "
             "have an argument on this space" in str(err.value))
     # Change to a space that is referenced by an operator
     code = code.replace("[W0, W3]", "[W0, W2]")
     ast = fpapi.parse(code, ignore_comments=False)
-    dkm = DynKernMetadata(ast, name="testkern_eval_type")
-    assert isinstance(dkm, DynKernMetadata)
+    dkm = LFRicKernMetadata(ast, name="testkern_eval_type")
+    assert isinstance(dkm, LFRicKernMetadata)
 
 
 def test_single_kern_eval(tmpdir):
@@ -1463,7 +1463,7 @@ def test_basis_evaluator():
 
     '''
     ast = fpapi.parse(BASIS_EVAL, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -1590,7 +1590,7 @@ def test_basis_unsupported_space():
     PSy layer to the kernels (see issue #461). '''
     # Test any_space_*
     ast = fpapi.parse(BASIS_UNSUPPORTED_SPACE, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     with pytest.raises(GenerationError) as excinfo:
@@ -1604,7 +1604,7 @@ def test_basis_unsupported_space():
                                            "any_discontinuous_space_5")
     code = code.replace("gh_inc", "gh_readwrite")
     ast = fpapi.parse(code, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     with pytest.raises(GenerationError) as excinfo:
@@ -1666,7 +1666,7 @@ def test_diff_basis():
 
     '''
     ast = fpapi.parse(DIFF_BASIS, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -1832,7 +1832,7 @@ def test_diff_basis_eval():
 
     '''
     ast = fpapi.parse(DIFF_BASIS_EVAL, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -1949,7 +1949,7 @@ def test_2eval_stubgen():
         "     integer :: gh_shape = gh_evaluator\n"
         "     integer :: gh_evaluator_targets(2) = (/w2h, wtheta/)\n")
     ast = fpapi.parse(twoeval_meta, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -2060,7 +2060,7 @@ def test_diff_basis_unsupp_space():
     kernels (see issue #461). '''
     # Test any_space_*
     ast = fpapi.parse(DIFF_BASIS_UNSUPPORTED_SPACE, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     with pytest.raises(GenerationError) as excinfo:
@@ -2074,7 +2074,7 @@ def test_diff_basis_unsupp_space():
                                                 "any_discontinuous_space_5")
     code = code.replace("gh_inc", "gh_readwrite")
     ast = fpapi.parse(code, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     with pytest.raises(GenerationError) as excinfo:
@@ -2089,7 +2089,7 @@ def test_dynbasisfns_unsupp_qr(monkeypatch):
     DynBasisFunctions._stub_declarations() if an un-supported quadrature
     shape is encountered. '''
     ast = fpapi.parse(DIFF_BASIS, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     dbasis = DynBasisFunctions(kernel)
@@ -2106,7 +2106,7 @@ def test_dynbasisfns_declns(monkeypatch):
     ''' Check the various internal errors that
     DynBasisFunctions._basis_fn_declns can raise. '''
     ast = fpapi.parse(DIFF_BASIS, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     dbasis = DynBasisFunctions(kernel)
