+++ src/psyclone/tests/dynamo0p3_stubgen_test.py
@@ -45,8 +45,7 @@ import fparser
 from fparser import api as fpapi
 
 from psyclone.configuration import Config
-from psyclone.domain.lfric import LFRicConstants, LFRicKern
-from psyclone.dynamo0p3 import DynKernMetadata
+from psyclone.domain.lfric import LFRicConstants, LFRicKern, LFRicKernMetadata
 from psyclone.errors import GenerationError
 from psyclone.gen_kernel_stub import generate
 
@@ -70,7 +69,7 @@ def test_kernel_stub_invalid_iteration_space():
     ast = fpapi.parse(os.path.join(BASE_PATH,
                                    "testkern_dofs_mod.f90"),
                       ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     with pytest.raises(GenerationError) as excinfo:
@@ -158,7 +157,7 @@ def test_load_meta_wrong_type():
     if the meta-data contains an un-recognised type. '''
     fparser.logging.disable(fparser.logging.CRITICAL)
     ast = fpapi.parse(INTENT, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     # Break the meta-data
     metadata.arg_descriptors[0]._argument_type = "gh_hedge"
@@ -172,7 +171,7 @@ def test_load_meta_wrong_type():
 def test_intent():
     ''' test that field intent is generated correctly for kernel stubs '''
     ast = fpapi.parse(INTENT, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = kernel.gen_stub
@@ -235,7 +234,7 @@ def test_spaces():
 
     '''
     ast = fpapi.parse(SPACES, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -348,7 +347,7 @@ def test_any_spaces():
 
     '''
     ast = fpapi.parse(ANY_SPACES, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -407,7 +406,7 @@ end module dummy_mod
 def test_vectors():
     ''' test that field vectors are handled correctly for kernel stubs '''
     ast = fpapi.parse(VECTORS, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = kernel.gen_stub
@@ -439,7 +438,7 @@ def test_arg_descriptor_vec_str():
     expected when we have a vector quantity. '''
     fparser.logging.disable(fparser.logging.CRITICAL)
     ast = fpapi.parse(VECTORS, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     field_descriptor = metadata.arg_descriptors[0]
     result = str(field_descriptor)
     expected_output = (
@@ -458,7 +457,7 @@ def test_enforce_bc_kernel_stub_gen():
     '''
     ast = fpapi.parse(os.path.join(BASE_PATH, "enforce_bc_kernel_mod.f90"),
                       ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = kernel.gen_stub
@@ -493,7 +492,7 @@ def test_enforce_op_bc_kernel_stub_gen():
     ast = fpapi.parse(os.path.join(BASE_PATH,
                                    "enforce_operator_bc_kernel_mod.F90"),
                       ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -526,7 +525,7 @@ def test_multi_qr_stub_gen():
     ast = fpapi.parse(os.path.join(BASE_PATH,
                                    "testkern_2qr_mod.F90"),
                       ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = str(kernel.gen_stub)
@@ -570,7 +569,7 @@ def test_qr_plus_eval_stub_gen():
     ast = fpapi.parse(os.path.join(BASE_PATH,
                                    "testkern_qr_eval_mod.F90"),
                       ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     gen_code = str(kernel.gen_stub)
@@ -628,7 +627,7 @@ def test_sub_name():
     name. In this case we append "_code to the name and _mod to the
     kernel name.'''
     ast = fpapi.parse(SUB_NAME, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kernel = LFRicKern()
     kernel.load_meta(metadata)
     generated_code = kernel.gen_stub
