+++ src/psyclone/tests/psyGen_test.py
@@ -52,30 +52,32 @@ from fparser.two import Fortran2003
 from psyclone.configuration import Config
 from psyclone.core.access_type import AccessType
 from psyclone.domain.common.psylayer import PSyLoop
-from psyclone.domain.lfric import LFRicKern, lfric_builtins
+from psyclone.domain.lfric import LFRicKern, lfric_builtins, LFRicKernMetadata
 from psyclone.domain.lfric.transformations import LFRicLoopFuseTrans
-from psyclone.dynamo0p3 import DynKernMetadata, DynInvokeSchedule, \
-    DynKernelArguments, DynGlobalSum
+from psyclone.dynamo0p3 import (DynInvokeSchedule, DynKernelArguments, 
+                                DynGlobalSum)
 from psyclone.errors import GenerationError, FieldNotFoundError, InternalError
 from psyclone.generator import generate
 from psyclone.gocean1p0 import GOKern
 from psyclone.parse.algorithm import parse, InvokeCall
-from psyclone.psyGen import TransInfo, Transformation, PSyFactory, \
-    InlinedKern, object_index, HaloExchange, Invoke, \
-    DataAccess, Kern, Arguments, CodedKern, Argument, GlobalSum, \
-    InvokeSchedule, BuiltIn
-from psyclone.psyir.nodes import Assignment, BinaryOperation, Container, \
-    Literal, Loop, Node, KernelSchedule, Call, colored
-from psyclone.psyir.symbols import DataSymbol, RoutineSymbol, REAL_TYPE, \
-    ImportInterface, ContainerSymbol, Symbol, INTEGER_TYPE, DeferredType, \
-    SymbolTable
+from psyclone.psyGen import (TransInfo, Transformation, PSyFactory,
+                             InlinedKern, object_index, HaloExchange, Invoke,
+                             DataAccess, Kern, Arguments, CodedKern, Argument,
+                             GlobalSum, InvokeSchedule, BuiltIn)
+from psyclone.psyir.nodes import (Assignment, BinaryOperation, Container,
+                                  Literal, Loop, Node, KernelSchedule, Call,
+                                  colored)
+from psyclone.psyir.symbols import (DataSymbol, RoutineSymbol, REAL_TYPE,
+                                    ImportInterface, ContainerSymbol, Symbol,
+                                    INTEGER_TYPE, DeferredType, SymbolTable)
 from psyclone.tests.lfric_build import LFRicBuild
 from psyclone.tests.test_files import dummy_transformations
 from psyclone.tests.test_files.dummy_transformations import LocalTransformation
 from psyclone.tests.utilities import get_invoke
-from psyclone.transformations import Dynamo0p3RedundantComputationTrans, \
-    Dynamo0p3KernelConstTrans, Dynamo0p3OMPLoopTrans, Dynamo0p3ColourTrans, \
-    OMPParallelTrans
+from psyclone.transformations import (Dynamo0p3RedundantComputationTrans,
+                                      Dynamo0p3KernelConstTrans, 
+                                      Dynamo0p3OMPLoopTrans, 
+                                      Dynamo0p3ColourTrans, OMPParallelTrans)
 
 
 BASE_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)),
@@ -474,7 +476,7 @@ def test_codedkern_node_str():
 
     '''
     ast = fpapi.parse(FAKE_KERNEL_METADATA, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     my_kern = LFRicKern()
     my_kern.load_meta(metadata)
     out = my_kern.node_str()
@@ -689,7 +691,7 @@ def test_kern_children_validation():
     '''
     # We use a subclass (CodedKern->LFRicKern) to test this functionality.
     ast = fpapi.parse(FAKE_KERNEL_METADATA, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     kern = LFRicKern()
     kern.load_meta(metadata)
 
@@ -779,7 +781,7 @@ def test_incremented_arg():
     # Therefore, we change the object produced by parsing the meta-data
     # instead
     ast = fpapi.parse(FAKE_KERNEL_METADATA, ignore_comments=False)
-    metadata = DynKernMetadata(ast)
+    metadata = LFRicKernMetadata(ast)
     for descriptor in metadata.arg_descriptors:
         if descriptor.access == AccessType.INC:
             descriptor._access = AccessType.READ
